Refactored Server setting APIs into a service to allow unit testing.
